+++ ctx.go
@@ -45,7 +45,6 @@ type contextKey int //nolint:unused // need for future (nolintlint)
 //
 //go:generate ifacemaker --file ctx.go --file req.go --file res.go --struct DefaultCtx --iface Ctx --pkg fiber --promoted --output ctx_interface_gen.go --not-exported true --iface-comment "Ctx represents the Context which hold the HTTP request and response.\nIt has methods for the request query string, parameters, body, HTTP headers and so on."
 type DefaultCtx struct {
-	userContext   context.Context      // User provided context
 	DefaultReq                         // Default request api
 	DefaultRes                         // Default response api
 	app           *App                 // Reference to *App
@@ -108,26 +107,23 @@ func (c *DefaultCtx) RequestCtx() *fasthttp.RequestCtx {
 	return c.fasthttp
 }
 
-// Context returns a standard context.Context containing all values stored in Locals
-// and any user-provided context set via SetContext. The returned context is safe
-// for use outside of the handler since it does not depend on the recycled Fiber Ctx.
+const userContextKey = "__fiber_context"
+
+// Context returns a context implementation that was set by
+// user earlier or returns a non-nil, empty context, if it was not set earlier.
 func (c *DefaultCtx) Context() context.Context {
-	ctx := c.userContext
-	if ctx == nil {
+	ctx, ok := c.fasthttp.UserValue(userContextKey).(context.Context)
+	if !ok {
 		ctx = context.Background()
+		c.SetContext(ctx)
 	}
-	c.fasthttp.VisitUserValuesAll(func(key, val any) {
-		ctx = context.WithValue(ctx, key, val)
-	})
+
 	return ctx
 }
 
-// SetContext sets a custom context.Context that will be returned by Context().
-// This can be used to propagate deadlines, cancelation signals, or values to
-// asynchronous operations started from the handler.
-func (c *DefaultCtx) SetContext(ctx context.Context) Ctx {
-	c.userContext = ctx
-	return c
+// SetContext sets a context implementation by user.
+func (c *DefaultCtx) SetContext(ctx context.Context) {
+	c.fasthttp.SetUserValue(userContextKey, ctx)
 }
 
 // Deadline returns the time when work done on behalf of this context
@@ -462,7 +458,7 @@ func (c *DefaultCtx) Reset(fctx *fasthttp.RequestCtx) {
 
 	c.DefaultReq.c = c
 	c.DefaultRes.c = c
-	c.userContext = nil
+	c.fasthttp.SetUserValue(userContextKey, nil)
 }
 
 // Release is a method to reset context fields when to use ReleaseCtx()
@@ -479,7 +475,6 @@ func (c *DefaultCtx) release() {
 		ReleaseRedirect(c.redirect)
 		c.redirect = nil
 	}
-	c.userContext = nil
 	c.DefaultReq.release()
 	c.DefaultRes.release()
 }